pyproject.toml
[project]
name = "adversarial-images"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.11.0",
    "datasets>=4.4.1",
    "jaxtyping>=0.3.3",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pyyaml>=6.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "transformers>=4.57.1",
    "typer>=0.9.0",
]
